refactor: split up code between dedicated files for better navigation
