+++ crates/logger/src/logger.rs
@@ -1,70 +1,16 @@
-use chrono::{DateTime, Local};
-use std::{fs::File,
-          io::{Write, IsTerminal},
-          sync::{atomic::{AtomicPtr, AtomicBool},
+use chrono::Local;
+use std::{sync::{atomic::{AtomicPtr, AtomicBool},
                  Arc
           },
           thread
 };
-use std::any::Any;
 use std::path::Path;
 use std::sync::atomic::AtomicU32;
-use std::thread::ThreadId;
-use crate::{LOGGER, LOGGING_QUEUE};
 
-#[derive(Clone)]
-pub struct LogMessage {
-    level: LogLevel,
-    thread_id: ThreadId,
-    timestamp: DateTime<Local>,
-    message: String,
-}
+use crate::{LOGGING_QUEUE};
+use crate::msg_fmt::*;
+use crate::targets::*;
 
-pub struct Colors {
-    info: String,
-    warn: String,
-    error: String,
-    debug: String,
-    trace: String
-}
-
-impl Colors {
-    fn new(message_fmt: String) -> Self {
-        Colors {
-            info: format!("\x1b[1;32m{}\x1b[0m", message_fmt), // green
-            warn: format!("\x1b[1;33m{}\x1b[0m", message_fmt), // yellow
-            error: format!("\x1b[1;31m{}\x1b[0m", message_fmt), // red
-            debug: format!("\x1b[1;34m{}\x1b[0m", message_fmt), // blue
-            trace: format!("\x1b[1;35m{}\x1b[0m", message_fmt) // magenta
-        }
-    }
-}
-
-impl std::fmt::Display for LogMessage {
-    fn fmt(&self,
-           f: &mut std::fmt::Formatter
-    ) -> std::fmt::Result {
-        let message_fmt = format!("{:?} - {} [{:5}] - {}",
-                                  self.thread_id,
-                                  self.timestamp.format("%d/%m/%Y %H:%M:%S.%f"),
-                                  format!("{:?}", self.level),
-                                  self.message
-        );
-        if std::io::stdout().is_terminal() {
-            let colors = Colors::new(message_fmt);
-            let colored_msg = match self.level {
-                LogLevel::Info => colors.info,
-                LogLevel::Warn => colors.warn,
-                LogLevel::Error => colors.error,
-                LogLevel::Debug => colors.debug,
-                LogLevel::Trace => colors.trace,
-            };
-            write!(f, "{}", colored_msg)
-        } else {
-            write!(f, "{}", message_fmt)
-        }
-    }
-}
 
 #[derive(Debug, Eq, PartialEq, PartialOrd, Ord, Clone, Copy)]
 pub enum LogLevel {
@@ -75,82 +21,10 @@ pub enum LogLevel {
     Trace,
 }
 
-pub trait LogTarget {
-    fn log(&self,
-           message: String
-    );
-    fn as_any(&self) -> &dyn Any;
-}
-
-pub struct NoopLogTarget;
-
-impl LogTarget for NoopLogTarget {
-    fn log(&self,
-           _message: String
-    ) {}
-
-    fn as_any(&self) -> &dyn Any {
-        self
-    }
-}
-
-pub struct ConsoleLogTarget;
-
-impl LogTarget for ConsoleLogTarget {
-    fn log(&self,
-           message: String
-    ) {
-        let result = writeln!(std::io::stdout(), "{}", message);
-        match result {
-            Ok(_) => {},
-            Err(_) => eprintln!("Failed to write to stdout! BAD!"),
-        }
-    }
-
-    fn as_any(&self) -> &dyn Any {
-        self
-    }
-}
-
-pub struct FileLogTarget {
-    file: File
-}
-
-impl LogTarget for FileLogTarget {
-    fn log(&self,
-           message: String
-    ) {
-        let result = writeln!(&self.file, "{}", message);
-        match result {
-            Ok(_) => {},
-            Err(_) => eprintln!("Failed to write to log file!"),
-        }
-    }
-
-    fn as_any(&self) -> &dyn Any {
-        self
-    }
-}
-
-impl FileLogTarget {
-    pub fn new(path: &Path) -> Self {
-        let file = std::fs::OpenOptions::new().append(true).create(true).open(path);
-        match file {
-            Ok(file) => FileLogTarget { file },
-            Err(err) => {
-                eprintln!("Failed to create file: {:?}! default file is generated!", err);
-                FileLogTarget {
-                    file: std::fs::OpenOptions::new().append(true).create(true).open("serverlog_default.txt").unwrap()
-                }
-            }
-        }
-    }
-}
-
 pub struct Logger {
     queue_capacity: Arc<AtomicU32>,
-    is_running: Arc<AtomicBool>,
-    severity_level: Arc<AtomicPtr<LogLevel>>,
+    pub(crate) is_running: Arc<AtomicBool>,
+    pub(crate) severity_level: Arc<AtomicPtr<LogLevel>>,
     targets: Arc<AtomicPtr<Vec<Box<dyn LogTarget + Send + Sync>>>>
 }
 
@@ -267,22 +141,5 @@ impl Logger {
         let level_ptr = self.severity_level.load(std::sync::atomic::Ordering::Acquire);
         let level = unsafe { &*level_ptr };
         *level
-    }
-}
-
-pub fn start_consumer_thread() {
-    let consumer_queue_ptr = LOGGING_QUEUE.clone();
-    unsafe {
-        thread::spawn(move || unsafe {
-            while LOGGER.is_running.load(std::sync::atomic::Ordering::Acquire)
-                || !consumer_queue_ptr.is_empty() {
-                while let Some(msg) = consumer_queue_ptr.pop() {
-                    if msg.level < *LOGGER.severity_level.load(std::sync::atomic::Ordering::Acquire) {
-                        continue;
-                    }
-                    LOGGER.log_write(&msg);
-                }
-            }
-        });
     }
 }
